+++ packages/brick-types/src/index.ts
@@ -2,3 +2,4 @@ export * from "./brick";
 export * from "./manifest";
 export * from "./menu";
 export * from "./runtime";
+export * from "./story";