+++ packages/babel-runtime-helpers/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/babel-runtime-helpers@0.1.2...@easyops/babel-runtime-helpers@0.1.3) (2020-07-29)
+
+**Note:** Version bump only for package @easyops/babel-runtime-helpers
+
+## [0.1.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/babel-runtime-helpers@0.1.1...@easyops/babel-runtime-helpers@0.1.2) (2020-07-21)
+
+**Note:** Version bump only for package @easyops/babel-runtime-helpers
+
 ## [0.1.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/babel-runtime-helpers@0.1.0...@easyops/babel-runtime-helpers@0.1.1) (2020-07-06)
 
 **Note:** Version bump only for package @easyops/babel-runtime-helpers